renamed to RecommendationRequest instead of Requests for folders
